+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add spinner for converting a single file.
 - Output names for concatenating filtered sequences is now default to the directory name with parameters as sufixes.
+- Allow "ignore" in the summary data types.
 
 ## v0.6.6